add useContext and handle state with it
