+++ code/src/App.js
@@ -1,13 +1,13 @@
 import React, { useState, useEffect } from "react";
 
+import FormContext from "./components/FormContext/FormContext.component";
 import MessageForm from "./components/NewMessageForm/NewMessageForm.component";
 import MessageList from "./components/MessageList/MessageList.component";
 
-import { API_URL, LIKES_URL } from "./api/urls";
+import { API_URL } from "./api/urls";
 
 export const App = () => {
   const [messageList, setMessageList] = useState([]);
-  const [newMessage, setNewMessage] = useState("");
 
   const fetchMessageList = () => {
     fetch(API_URL)
@@ -19,54 +19,10 @@ export const App = () => {
     fetchMessageList();
   }, []);
 
-  const handleNewMessageChange = (e) => {
-    setNewMessage(e.target.value);
-  };
-
-  const handleFormSubmit = (e) => {
-    e.preventDefault();
-
-    const options = {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: JSON.stringify({ text: newMessage }),
-    };
-
-    fetch(API_URL, options)
-      .then((res) => res.json())
-      .then((receivedMessage) =>
-        setMessageList([receivedMessage, ...messageList])
-      )
-      .catch((err) => console.error(err));
-  };
-
-  const handleLikesIncrease = (id) => {
-    const options = {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-    };
-
-    fetch(LIKES_URL(id), options)
-      .then((res) => res.json())
-      .then(() => fetchMessageList())
-      .catch((err) => console.error(err));
-  };
-
   return (
-    <>
-      <MessageForm
-        newMessage={newMessage}
-        handleNewMessageChange={handleNewMessageChange}
-        handleFormSubmit={handleFormSubmit}
-      />
-      <MessageList
-        messageList={messageList}
-        handleLikesIncrease={handleLikesIncrease}
-      />
-    </>
+    <FormContext.Provider value={{ messageList, setMessageList }}>
+      <MessageForm />
+      <MessageList />
+    </FormContext.Provider>
   );
 };
