+++ code/src/components/FormContext/FormContext.component.jsx
@@ -1,0 +1,3 @@
+import { createContext } from "react";
+
+export default createContext(null);
