+++ code/src/components/MessageList/MessageList.component.jsx
@@ -1,16 +1,19 @@
-import React from "react";
+import React, { useContext } from "react";
 
+import FormContext from '../FormContext/FormContext.component'
 import MessageListItem from "../MessageListItem/MessageListItem.component";
 
 import {MessageListContainer, MessageListItemContainer} from './MessageList.style'
 
-const MessageList = ({ messageList, handleLikesIncrease }) => {
+const MessageList = () => {
+  const { messageList } = useContext(FormContext);
+  
   return (
     <MessageListContainer>
       {messageList.map((message) => {
         return (
           <MessageListItemContainer key={message._id}>
-            <MessageListItem message={message} handleLikesIncrease={handleLikesIncrease} /> 
+            <MessageListItem message={message} /> 
           </MessageListItemContainer>
         )
       })}
